+++ reactive-banana/src/Reactive/Banana/Prim/Types.hs
@@ -10,7 +10,8 @@ import           Control.Monad.Trans.Reader
 import           Control.Monad.Trans.ReaderWriterIO
 import           Data.Functor
 import           Data.Hashable
-import           Data.Monoid
+import           Data.Monoid (Monoid, mempty, mappend)
+import           Data.Semigroup
 import qualified Data.Vault.Lazy                    as Lazy
 import           System.IO.Unsafe
 import           System.Mem.Weak
@@ -33,6 +34,7 @@ type Inputs        = ([SomeNode], Lazy.Vault)
 type EvalNetwork a = Network -> IO (a, Network)
 type Step          = EvalNetwork (IO ())
 
+emptyNetwork :: Network
 emptyNetwork = Network
     { nTime    = next beginning
     , nOutputs = OB.empty
@@ -51,9 +53,12 @@ newtype BuildW = BuildW (DependencyBuilder, [Output], Action, Maybe (Build ()))
     --          , late build actions
     --          )
 
+instance Semigroup BuildW where
+    BuildW x <> BuildW y = BuildW (x <> y)
+
 instance Monoid BuildW where
-    mempty                          = BuildW mempty
-    (BuildW x) `mappend` (BuildW y) = BuildW (x `mappend` y)
+    mempty  = BuildW mempty
+    mappend = (<>)
 
 type BuildIO = Build
 
@@ -70,13 +75,19 @@ ground = 0
 
 -- | 'IO' actions as a monoid with respect to sequencing.
 newtype Action = Action { doit :: IO () }
+instance Semigroup Action where
+    Action x <> Action y = Action (x >> y)
 instance Monoid Action where
     mempty = Action $ return ()
-    (Action x) `mappend` (Action y) = Action (x >> y)
+    mappend = (<>)
 
 -- | Lens-like functionality.
 data Lens s a = Lens (s -> a) (a -> s -> s)
-set    (Lens _   set)   = set
+
+set :: Lens s a -> a -> s -> s
+set (Lens _   set)   = set
+
+update :: Lens s a -> (a -> a) -> s -> s
 update (Lens get set) f = \s -> set (f $ get s) s
 
 {-----------------------------------------------------------------------------
@@ -128,6 +139,7 @@ instance Eq SomeNode where
     (P x) == (P y) = equalRef x y
     (L x) == (L y) = equalRef x y
     (O x) == (O y) = equalRef x y
+    _     == _     = False
 
 {-# INLINE mkWeakNodeValue #-}
 mkWeakNodeValue :: SomeNode -> v -> IO (Weak v)
@@ -136,11 +148,22 @@ mkWeakNodeValue (L x) = mkWeakRefValue x
 mkWeakNodeValue (O x) = mkWeakRefValue x
 
 -- Lenses for various parameters
-seenP  = Lens _seenP  (\a s -> s { _seenP = a })
-seenL  = Lens _seenL  (\a s -> s { _seenL = a })
+seenP :: Lens (Pulse' a) Time
+seenP = Lens _seenP  (\a s -> s { _seenP = a })
+
+seenL :: Lens (Latch' a) Time
+seenL = Lens _seenL  (\a s -> s { _seenL = a })
+
+valueL :: Lens (Latch' a) a
 valueL = Lens _valueL (\a s -> s { _valueL = a })
-parentsP  = Lens _parentsP (\a s -> s { _parentsP = a })
+
+parentsP :: Lens (Pulse' a) [Weak SomeNode]
+parentsP = Lens _parentsP (\a s -> s { _parentsP = a })
+
+childrenP :: Lens (Pulse' a) [Weak SomeNode]
 childrenP = Lens _childrenP (\a s -> s { _childrenP = a })
+
+levelP :: Lens (Pulse' a) Int
 levelP = Lens _levelP (\a s -> s { _levelP = a })
 
 -- | Evaluation monads.
@@ -185,9 +208,12 @@ beginning = T 0
 next :: Time -> Time
 next (T n) = T (n+1)
 
+instance Semigroup Time where
+    T x <> T y = T (max x y)
+
 instance Monoid Time where
-    mappend (T x) (T y) = T (max x y)
-    mempty              = beginning
+    mappend = (<>)
+    mempty  = beginning
 
 {-----------------------------------------------------------------------------
     Notes